Move helpers to utils.js, remove todos
